BT: Do not send hci raw command when stack has not started
Problem: Crash occurs when hci raw command is sent by apps
when Bluetooth is being initialized.

Reason: The check for stack status is not present in api to
send raw hci commands from btif. This leads to call reaching
hci layer even when Bluetooth is not fully enabled. At hci
layer LPM is accessed to post wake deassert on hci thread.
Accessing hci thread when it is not running and yet to be
initialized results in crash.

Fix: The btif api to send raw command must be guarded with
check for Bluetooth state. Added check to return error from
the api if Bluetooth stack has not stared.

Change-Id: I777231f224015fcbc13b78f79008a684af7cf4d8
CRs-Fixed: 1117351